--- src/main/java/Task.java
@@ -1,32 +1,0 @@
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "[X]" : "[ ]"); // mark done task with X
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void unmark() {
-         isDone = false;
-    }
-
-    @Override
-    public String toString() {
-       return getStatusIcon() + " " + description;
-    }
-
-    public String toSaveString() {
-        return getStatusIcon() + " | " + description;
-    }
-
-}
-
