+++ src/main/java/rob/DukeException.java
@@ -1,3 +1,5 @@
+package rob;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
