+++ src/main/java/rob/Event.java
@@ -1,4 +1,4 @@
-import java.io.StringReader;
+package rob;
 
 public class Event extends Task {
