+++ src/main/java/rob/Parser.java
@@ -1,3 +1,5 @@
+package rob;
+
 import java.util.Objects;
 
 public class Parser {
