+++ src/main/java/rob/Rob.java
@@ -1,3 +1,5 @@
+package rob;
+
 import java.io.IOException;
 import java.util.Objects;
 import java.util.Scanner;
