+++ src/main/java/rob/Storage.java
@@ -1,3 +1,5 @@
+package rob;
+import rob.*;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
